fix: code refactoring (#2)
refactoring
